pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "check_me"
version = "0.1.0"
description = "C/C++-first deterministic static security candidate generator"
requires-python = ">=3.10"
dependencies = [
    "click>=8.1",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
    "httpx>=0.27",
]

[project.scripts]
check_me = "check_me.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"
